refactor: move C-arm controls to separate module
